created several tests

+++ oop_java/Stack_Calculator/Stack_Calculator/src/main/java/calculator/context/ExecutableContex.java
@@ -41,7 +41,7 @@ public class ExecutableContex {
         return value;
     }
 
-    public double size() {
+    public int size() {
         return stack.size();
     }
